validated more API's are using source serializer correctly for writing
