+++ src/Tests/Framework/MockData/Project.cs
@@ -27,6 +27,8 @@ namespace Tests.Framework.MockData
 		public CompletionField Suggest { get; set; }
 		public IEnumerable<string> Branches { get; set; }
 		public Ranges Ranges { get; set; }
+		[JsonIgnore]
+		public string UniqueToSourceSerializer { get; set; }
 
 		public static Faker<Project> Generator { get; } =
 			new Faker<Project>()
